Rename the file-based CMS service to make it more intuitive
